requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
python-dotenv
python-jose
passlib[bcrypt]
python-multipart
